+++ packages/ical-tools/src/lib/vevent.ts
@@ -55,6 +55,7 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
   readonly start: InstanceType<T>;
   readonly isInfinite: boolean;
   readonly duration?: number | InstanceType<T>;
+  readonly maxDuration!: number;
   readonly hasDuration: boolean;
   readonly timezone: string | null;
 
@@ -240,6 +241,10 @@ export class VEvent<T extends typeof DateAdapter, D = any> extends OccurrenceGen
     }
 
     this.hasDuration = !!this._duration;
+
+    if (this._duration) {
+      this.maxDuration = this._duration;
+    }
 
     if (options.rrules) {
       this.rrules = options.rrules.map(ruleArgs => {